requirements.txt
cassandra-driver~=3.25.0
setuptools==65.4.1